feat: Clean installation custom special folders (#20)
A new argument, `cleanCustomSpecialFolders`, has been defined to enable the deletion of directories generated and utilized by a GeneXus installation. This feature is crucial for managing and preventing the unchecked accumulation of these directories, especially on non-ephemeral machines. To ensure backward compatibility with existing workflows, these directories will continue to be preserved by default; explicit action is required to clear their content.